requirements.txt
google-generativeai
streamlit
Pillow